+++ api_v3/lib/reflection/KalturaServicesMap.php
@@ -127,6 +127,7 @@ class KalturaServicesMap
 			$serviceMapEntry->serviceId = $serviceId;
 			$serviceMapEntry->serviceClass = $serviceClass;
 			$serviceMapEntry->serviceInfo = $serviceReflectionClass->getServiceInfo();
+
             $actionMap = array();
             $nativeActions = $serviceReflectionClass->getActions();
             foreach ($nativeActions as $actionId => $actionName)