+++ deployment/base/sql/01.kaltura_ce_tables.sql
@@ -53,9 +53,9 @@ CREATE TABLE IF NOT EXISTS app_token(
   `kuser_id` int(11) DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `int_id_index` (`int_id`),
-  KEY `kuser_id` (`kuser_id`)
+  KEY `kuser_id` (`kuser_id`),
   KEY `partner_id_status` (`partner_id`,`status`)
-)ENGINE=INNODB DEFAULT CHARSET=utf8;
+) ENGINE=INNODB DEFAULT CHARSET=utf8;
 
 CREATE TABLE IF NOT EXISTS `audit_trail` (
   `id` int(11) NOT NULL AUTO_INCREMENT,