+++ plugins/admin_console/system_partner/lib/api/filter/KalturaSystemPartnerFilter.php
@@ -9,11 +9,20 @@ class KalturaSystemPartnerFilter extends KalturaPartnerFilter
 	(
 		"partnerParentIdEqual" => "_eq_partner_parent_id",
 		"partnerParentIdIn" => "_in_partner_parent_id",
+		"adminEmailEqual" => "_eq_admin_email"
 	);
 	
 	public function getMapBetweenObjects()
 	{
 		return array_merge(parent::getMapBetweenObjects(), self::$map_between_objects);
+	}
+
+	/* (non-PHPdoc)
+	 * @see KalturaFilter::getCoreFilter()
+	 */
+	protected function getCoreFilter()
+	{
+		return new systemPartnerFilter();
 	}
 	
 	/**
@@ -27,4 +36,16 @@ class KalturaSystemPartnerFilter extends KalturaPartnerFilter
 	 * @requiresPermission all
 	 */
 	public $partnerParentIdIn;
+
+	/**
+	 * @var string
+	 * @requiresPermission all
+	 */
+	public $adminEmailEqual;
+
+	public function toObject ( $object_to_fill = null, $props_to_skip = array() )
+	{
+		$this->adminEmailEqual = trim($this->adminEmailEqual);
+		return parent::toObject($object_to_fill,$props_to_skip);
+	}
 }